pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents-builder>=0.1.4",
    "strands-agents-tools>=0.1.7",
    "duckduckgo-search>=5.3.1b1",
]

[tool.setuptools]
packages = ["deep_research_agent"]
